feat: add menu for AdventureGame
