+++ src/components/Sidebar/sui_menu.ts
@@ -1,5 +1,6 @@
 // import Games from "@mui/icons-material/Games";
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
+import SportsEsportsIcon from '@mui/icons-material/SportsEsports';
 
 const suiMenu = [
     {
@@ -7,6 +8,12 @@ const suiMenu = [
         name_i18_key: "menu.sui_constract_call",
         icon: PlayArrowIcon,
         path: "/sui/contract_call"
+    },
+    {
+        name: "sui_adventure_game",
+        name_i18_key: "menu.sui_adventure_game",
+        icon: SportsEsportsIcon,
+        path: "/sui/adventure_game"
     }
 ]
