+++ src/router/router.tsx
@@ -28,7 +28,8 @@ import ContractCall from "../pages/Constract/Call/index";
 import ABIParser from "../pages/Aptos/ABIParser";
 import MoveU8toString from "../pages/Aptos/MoveU8toString";
 
-import SuiConstractCall from "../pages/Sui/ConstractCall"
+import SuiConstractCall from "../pages/Sui/ConstractCall";
+import SuiAdventureGame from "../pages/Sui/AdventureGame"
 
 export default function Router() {
     return (<Routes>
@@ -62,6 +63,7 @@ export default function Router() {
         <Route path="/aptos/debug_string" element={<MoveU8toString />} />
 
         <Route path="/sui/contract_call" element={<SuiConstractCall />} />
+        <Route path="/sui/adventure_game" element={<SuiAdventureGame />} />
 
     </Routes>);
 }
